preprocessing/requirements.txt
# 预处理模块依赖项
chardet>=4.0.0
Pillow>=8.0.0
PyPDF2>=3.0.0
pdf2image>=1.16.0
pandas>=1.3.0
openai>=1.0.0
pydub>=0.25.0
python-docx>=0.8.0